==> RQ2/0-Programmer reuse statistics.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re


path1 = 'D:/CCFinder/Commit2So(A commit will correspond to multiple sos)/'
commit1_list = os.listdir(path1)
print(len(commit1_list))
commit1_list[:5]

path2 = 'D:/CCFinder/Commit2So(A commit only corresponds to the latest so)/'
commit2_list = os.listdir(path2)
print(len(commit2_list))
commit2_list[:5]

for file in tqdm(commit1_list):
    with open(path1+file, 'r', encoding='utf-8') as f1:
        lines = f1.readlines()
        latest_date = pd.to_datetime('1-1-1')
        for so_post in lines:
            CreationDate = re.findall(r'CreationDate = (.*?)//', so_post)[0][:10]
            if pd.to_datetime(CreationDate)>latest_date:
                latest_date = pd.to_datetime(CreationDate)
                res = so_post
        with open(path2+file, 'w', encoding='utf-8') as f2:
            f2.write(res)


lines

res
