==> RQ2/1-Programmer reuse statistics.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd

df1 = pd.read_excel('./Programmer Commit.xlsx')
df1.info()

commit_1 = df1[(df1['commit counts']==1)]
commit_1 = commit_1[commit_1['reuse counts']!=0]
print(len(commit_1))

commit_2 = df1[(df1['commit counts']==2)]
commit_2 = commit_2[commit_2['reuse counts']!=0]
print(len(commit_2))

commit_3t5 = df1[(df1['commit counts']>=3) & (df1['commit counts']<=5)]
commit_3t5 = commit_3t5[commit_3t5['reuse counts']!=0]
print(len(commit_3t5))

commit_6t15 = df1[(df1['commit counts']>=6) & (df1['commit counts']<=15)]
commit_6t15 = commit_6t15[commit_6t15['reuse counts']!=0]
print(len(commit_6t15))

commit_16t100 = df1[(df1['commit counts']>=16) & (df1['commit counts']<=100)]
commit_16t100 = commit_16t100[commit_16t100['reuse counts']!=0]
print(len(commit_16t100))


commit_100 = df1[(df1['commit counts']>100)]
commit_100  = commit_100 [commit_100['reuse counts']!=0]
print(len(commit_100 ))

df2 = pd.read_csv('../all_commit_info.csv',encoding='gbk')
df2 = df2[df2['Number of hits on SO']!=0]
df2.info()

commit_gp = [commit_1, commit_2, commit_3t5, commit_6t15, commit_16t100, commit_100]
path_dict = {0:'./1/', 1:'./2/', 2:'./3-5/',
             3:'./6-15/', 4:'./16-100/', 5:'./100/'}
for idx, co in enumerate(commit_gp):
    names = co['Commiter'].to_list()
    save_info = []
    for n in names:
        if n in df2['Commiter'].to_list():
            save_info.extend(df2[df2['Commiter']==n]['Modified snippets involved in ccfinder'].to_list())
    print(len(save_info))
    with open(path_dict[idx]+'commit info.txt','w',encoding='utf-8') as f:
        for line in save_info:
            f.write(line)
            f.write('\n')
