==> RQ2/2-Programmer reuse statistics.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd



path_dict = {0:'./1/', 1:'./2/', 2:'./3-5/',
             3:'./6-15/', 4:'./16-100/', 5:'./100/'}

for idx in tqdm(range(6)):
    with open(path_dict[idx]+'commit info.txt') as f1:
        lines = f1.readlines()
        with open(path_dict[idx]+'so info.txt','w',encoding='utf-8') as f2:
            for l in lines:
                l = l.rstrip().replace(':/','/').replace('/','-')
                with open('D:/CCFinder/Commit2So(A commit only corresponds to the latest so)/'+l+'.txt', encoding='utf-8') as f3:
                    solines = f3.readlines()
                    for sol in solines:
                        f2.write(sol)


lines[-1].rstrip().replace(':/','/').replace('/','-')


for idx in tqdm(range(6)):
    with open(path_dict[idx]+'so info.txt',encoding='utf-8') as f:
        lines = f.readlines()
        print(len(lines),len(set(lines)))

df2 = pd.read_csv('../all_commit_info.csv',encoding='gbk')

df2.info()
