==> RQ2/3-Programmer reuse statistics.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re

def extract_info(x):
    CreationDate = re.findall(r'CreationDate = (.*?)//',x)[0]

    Title = re.findall(r'Title = (.*?)//',x)[0]

    Tags = re.findall(r'Tags = (.*?)//',x)[0]
    Tags = Tags.split('><')
    Tags[0] = Tags[0].lstrip('<')
    Tags[-1] = Tags[-1].rstrip('>')
    Tags = ' '.join(Tags)

    Score = re.findall(r'Score = (.*?)//',x)[0]


    AnswerCount = re.findall(r'AnswerCount = (.*?)//',x)
    if AnswerCount==[] or AnswerCount==['']:
        AnswerCount = re.findall(r'AnswerCount = (.*?)\n',x)
        if AnswerCount == [] or AnswerCount==['']:
            AnswerCount = None
        else:
            AnswerCount = AnswerCount[0]
    else:
        AnswerCount = AnswerCount[0]


    CommentCount = re.findall(r'CommentCount = (.*?)//',x)
    if CommentCount==[] or CommentCount==['']:
        CommentCount = re.findall(r'CommentCount = (.*?)\n',x)
        if CommentCount==[] or CommentCount==['']:
            CommentCount = None
        else:
            CommentCount = CommentCount[0]
    else:
        CommentCount = CommentCount[0]

    FavoriteCount = re.findall(r'FavoriteCount = (.*?)//',x)
    if FavoriteCount==[] or FavoriteCount==['']:
        FavoriteCount = re.findall(r'FavoriteCount = (.*?)\n',x)
        if FavoriteCount==[] or FavoriteCount==['']:
            FavoriteCount = None
        else:
            FavoriteCount = FavoriteCount[0]
    else:
        FavoriteCount = FavoriteCount[0]



    return CreationDate, Title, Tags, Score, AnswerCount, CommentCount, FavoriteCount


path_dict = {0:'./1/', 1:'./2/', 2:'./3-5/',
             3:'./6-15/', 4:'./16-100/', 5:'./100/'}

for idx in tqdm(range(6)):
    new_df = pd.DataFrame({'CreationDate':[],'Title':[], 'Tags':[],
                              'Score':[],'AnswerCount':[], 'CommentCount':[],
                              'FavoriteCount':[]})
    with open(path_dict[idx]+'so info.txt', encoding='utf-8') as f:
        lines = f.readlines()
        lines = list(set(lines))

        for l in lines:
            CreationDate, Title, Tags, Score, AnswerCount, CommentCount, FavoriteCount = extract_info(l)
            new_df.loc[len(new_df)] = [CreationDate, Title, Tags, Score, AnswerCount, CommentCount, FavoriteCount]


    new_df.to_excel(path_dict[idx]+'so details.xlsx',index=False)
