==> RQ2/5-Programmer reuse statistics(Tags).ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt

path_dict = {0:'./1/', 1:'./2/', 2:'./3-5/',
             3:'./6-15/', 4:'./16-100/', 5:'./100/'}


df_list = []
for i in range(6):
    df = pd.read_excel(path_dict[i]+'so details.xlsx')
    print(df.info())
    df_list.append(df)


df1, df2, df3, df4, df5, df6 = df_list


df1['Tags'][0].split(' ')
Tags = [[],[],[],[],[],[]]
for idx, temp_df in enumerate(df_list):
    for i in range(len(temp_df)):
        Tags[idx].extend(temp_df['Tags'][i].split(' '))



tag_dict_list = [dict(),dict(),dict(),dict(),dict(),dict()]
for idx, tag in enumerate(Tags):
    for t in tag:
        if t in tag_dict_list[idx].keys():
            tag_dict_list[idx][t] += 1
        else:
            tag_dict_list[idx][t] = 0




tag_dict_list[1]


tag_dict1, tag_dict2, tag_dict3, tag_dict4, tag_dict5, tag_dict6 = tag_dict_list


sorted(tag_dict1.items(), key=lambda x: x[1],reverse=True)[1:6]


sorted(tag_dict2.items(), key=lambda x: x[1],reverse=True)[1:6]

sorted(tag_dict3.items(), key=lambda x: x[1],reverse=True)[1:6]


sorted(tag_dict4.items(), key=lambda x: x[1],reverse=True)[1:6]


sorted(tag_dict5.items(), key=lambda x: x[1],reverse=True)[1:6]


sorted(tag_dict6.items(), key=lambda x: x[1],reverse=True)[1:6]
